requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
flake8==3.8.4
httpie==2.4.0
idna==2.10
iniconfig==1.1.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
pep8==1.7.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
pylint==2.6.2
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
